requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langgraph==0.2.20
langchain-openai==0.2.0
langchain-google-genai==2.0.0
pypdf
faiss-cpu
streamlit==1.38.0
python-dotenv

